getx, nav2를 분리하여 먼저 getx로 처리. encrypt, decrypt 처리
