+++ lib/config.dart
@@ -1,0 +1,4 @@
+class AppConfig {
+  static String key = 'Z9HF6BED46L1D3O3C3DEBE8U26T74FNY';
+  static String iv = 'hEl0loF1aCt3oRy3';
+}
